api/requirements.txt
# FastAPI Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
asyncpg==0.29.0

# Data Processing
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
prophet==1.1.5
statsmodels==0.14.1
joblib==1.3.2
tqdm==4.66.1
xgboost==2.0.3
shap==0.45.0

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
